pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "glpi-dashboard-cau"
version = "0.1.0"
description = "Service desk dashboard for GLPI"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "dash>=2.17",
    "pandas>=2.2",
    "requests>=2.32",
    "python-dotenv>=1.0",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "strawberry-graphql>=0.223",
    "httpx>=0.27",
    "rich-click>=1.6",
    "aiohttp",
    "aiohttp_retry",
    "APScheduler",
    "SQLAlchemy==2.0.30",
    "asyncpg",
    "redis",
    "Flask-Caching>=2.1",
    "flask-compress",
    "langgraph>=0.0.22",
    "opentelemetry-api>=1.24",
    "opentelemetry-sdk>=1.24",
    "opentelemetry-exporter-otlp-proto-http>=1.24",
    "opentelemetry-instrumentation-fastapi>=0.55b1",
    "opentelemetry-instrumentation-logging>=0.55b1",
    "gql[httpx]>=3.5",
    "dishka>=1.6",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "tenacity>=8.2",
    "purgatory>=3.0",
    "loguru>=0.7",
    "orjson>=3.10",
    "arq>=0.26",
    "pybreaker>=0.20",
    "prometheus-client",
    "prometheus-fastapi-instrumentator>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.1",
    "pytest-cov==6.2.1",
    "pytest-asyncio==1.0.0",
    "black",
    "mypy",
    "ruff",
    "flake8",
    "isort",
    "pytest-mock==3.14.1",
    "pytest-httpx==0.30.0",
    "testcontainers",
    "playwright",
    "pact-python",
    "types-requests",
    "types-redis",
    "pydantic-to-typescript",
    "rope",
]

[project.scripts]
glpi-worker = "worker:main"
glpi-dashboard = "dashboard_app:main"
glpi-tools = "glpi_tools.__main__:cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
check_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.pdm]
package-type = "application"

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["app*", "tests*", "scripts*", "examples*", "labs*"]
